move repository classes for splatoon 2 into an own package
