+++ src/main/java/tv/strohi/twitch/strohkoenigbot/splatoonapi/rotations/StagesExporter.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import tv.strohi.twitch.strohkoenigbot.chatbot.spring.DiscordBot;
 import tv.strohi.twitch.strohkoenigbot.data.model.splatoon2.splatoondata.SplatoonStage;
-import tv.strohi.twitch.strohkoenigbot.data.repository.splatoondata.SplatoonStageRepository;
+import tv.strohi.twitch.strohkoenigbot.data.repository.splatoon2.splatoondata.SplatoonStageRepository;
 import tv.strohi.twitch.strohkoenigbot.splatoonapi.model.SplatNetStage;
 import tv.strohi.twitch.strohkoenigbot.utils.DiscordChannelDecisionMaker;
